requirement.txt
absl-py==0.11.0
adal
aenum==3.0.0
aiohttp
appdirs==1.4.4
applicationinsights==0.11.9
argcomplete==1.12.2
argon2-cffi==20.1.0
astroid==2.5
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
azure-cli-core
azure-cli-telemetry==1.0.6
azure-common==1.1.26
azure-core==1.11.0
azure-mgmt-core==1.2.2
azure-nspkg==3.0.2
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-storage-nspkg==3.1.0
backcall==0.2.0
backports.lzma==0.0.14
bcrypt==3.2.0
bidict==0.21.2
bleach==3.3.0
boto==2.49.0
boto3==1.17.13
botocore==1.20.13
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
control==0.8.4
coverage==5.4
crcmod==1.7
cryptography==3.4.6
cycler==0.10.0
cysignals==1.10.2
Cython==0.29.22
datadog==0.40.0
decorator==4.4.2
defusedxml==0.6.0
dictdiffer==0.8.1
distlib==0.3.1
elasticsearch==7.11.0
et-xmlfile==1.0.1
fastcluster==1.1.28
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
flatbuffers==1.12
gast==0.3.3
GeoAlchemy2==0.8.4
git-pylint-commit-hook==2.5.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.9.0
hexdump==3.3
humanfriendly==9.1
identify==1.5.14
idna==2.10
imageio==2.9.0
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
json-logging-py==0.2
json-rpc==1.13.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
knack==0.8.0rc2
lazy-object-proxy==1.5.2
libusb1==1.9.2
Logentries @ git+https://github.com/commaai/le_python.git@feaeacb48f7f4bdb02c0a8fc092326d4e101b7f2
lru-dict==1.1.7
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
mpld3==0.5.2
mpmath==1.2.1
msal==1.0.0
msal-extensions==0.1.3
msgpack-python==0.5.6
msrest==0.6.21
msrestazure==0.6.4
multidict==5.1.0
mypy==0.812
mypy-extensions==0.4.3
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
nodeenv==1.5.0
nose==1.3.7
notebook==6.2.0
numpy==1.20.1
onnx==1.8.1
onnxruntime==1.6.0
opencv-python==4.5.1.48
openpyxl==3.0.6
opt-einsum==3.3.0
osmium==3.1.3
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parameterized==0.8.1
paramiko==2.7.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pipenv==2020.8.13
pkginfo==1.7.0
portalocker==1.7.1
pprofile==2.0.5
pre-commit==2.10.1
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.1
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycapnp==1.0.0b1
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
pycurl==7.43.0.6
pyflakes==2.2.0
pygame==2.0.0.dev8
Pygments==2.8.0
PyJWT==2.0.1
pylint==2.7.0
pymongo==3.11.3
PyMySQL==0.10.1
PyNaCl==1.4.0
pynmea2==1.16.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyprof2calltree==1.4.5
pyproj==3.0.0.post1
pyrsistent==0.17.3
pyserial==3.5
python-dateutil==2.8.1
python-engineio==4.0.0
python-logstash==0.4.6
python-socketio==5.0.4
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
raven==6.10.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
reverse-geocoder==1.5.1
rsa==4.6
s2sphere==0.2.5
s3transfer==0.3.4
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
SCons==4.1.0.post1
seaborn==0.11.1
Send2Trash==1.5.0
setproctitle==1.2.2
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smbus2==0.4.1
SQLAlchemy==1.3.23
subprocess32==3.5.4
sympy==1.7.1
tabulate==0.8.9
tenacity==6.3.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.2.1
timezonefinder==5.2.0
toml==0.10.2
tornado==6.1
tqdm==4.57.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
utm==0.7.0
virtualenv==20.4.2
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zmq==0.0.0
